backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]>=3.2.2,<3.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
httpx==0.27.2
email-validator==2.2.0
alembic==1.13.1
pytest==8.2.2
numpy==1.26.4
prometheus_client==0.20.0
locust==2.24.0
PyJWT==2.8.0
aiosqlite==0.20.0
asyncpg==0.29.0
